Moved some VCS interfaces out of coreApi and into versionControl project.
