+++ subprojects/version-control/src/main/java/org/gradle/vcs/internal/VersionControlSpecFactory.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2017 the original author or authors.
+ * Copyright 2018 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,27 +16,9 @@
 
 package org.gradle.vcs.internal;
 
-import org.gradle.StartParameter;
 import org.gradle.api.internal.initialization.ClassLoaderScope;
-import org.gradle.internal.Cast;
-import org.gradle.internal.reflect.Instantiator;
 import org.gradle.vcs.VersionControlSpec;
-import org.gradle.vcs.git.GitVersionControlSpec;
-import org.gradle.vcs.git.internal.DefaultGitVersionControlSpec;
 
-public class VersionControlSpecFactory {
-    private final Instantiator instantiator;
-    private final StartParameter rootBuildStartParameter;
-
-    public VersionControlSpecFactory(Instantiator instantiator, StartParameter rootBuildStartParameter) {
-        this.instantiator = instantiator;
-        this.rootBuildStartParameter = rootBuildStartParameter;
-    }
-
-    public <T extends VersionControlSpec> T create(Class<T> specType, ClassLoaderScope classLoaderScope) {
-        if (specType.isAssignableFrom(GitVersionControlSpec.class)) {
-            return Cast.uncheckedCast(instantiator.newInstance(DefaultGitVersionControlSpec.class, rootBuildStartParameter, classLoaderScope));
-        }
-        throw new IllegalArgumentException(String.format("Do not know how to create an instance of %s.", specType.getName()));
-    }
+public interface VersionControlSpecFactory {
+    <T extends VersionControlSpec> T create(Class<T> specType, ClassLoaderScope classLoaderScope);
 }
