+++ subprojects/version-control/src/main/java/org/gradle/vcs/internal/VersionRef.java
